+++ src/components/MapControls/MapControls.js
@@ -1,11 +1,9 @@
 /* eslint-disable no-underscore-dangle */
-import React, { useEffect, useMemo, useState } from "react";
+import React, { useMemo } from "react";
 import PropTypes from "prop-types";
 import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
 import Zoom from "react-spatial/components/Zoom";
-import { ZoomSlider } from "ol/control";
-import { unByKey } from "ol/Observable";
 import { makeStyles } from "@mui/styles";
 import FitExtent from "./FitExtent";
 import Geolocation from "./Geolocation";
@@ -15,12 +13,13 @@ import { ReactComponent as ZoomOut } from "../../img/minus.svg";
 import { ReactComponent as ZoomIn } from "../../img/plus.svg";
 import useHasScreenSize from "../../utils/useHasScreenSize";
 import { setZoomType } from "../../model/map/actions";
+import FloorSwitcher from "../FloorSwitcher";
 
 const propTypes = {
   geolocation: PropTypes.bool,
-  zoomSlider: PropTypes.bool,
   fitExtent: PropTypes.bool,
   menuToggler: PropTypes.bool,
+  floorSwitcher: PropTypes.bool,
   children: PropTypes.node,
 };
 
@@ -39,41 +38,45 @@ const useStyles = makeStyles((theme) => {
         position: "relative",
         display: "flex",
         flexDirection: "column",
-        gap: (props) => {
-          return props.margin;
+        boxShadow: "0 0 7px rgb(0 0 0 / 90%)",
+        transition: "box-shadow 0.5s ease",
+        borderRadius: 20,
+        gap: 1,
+        "&::after": {
+          position: "absolute",
+          content: '""',
+          top: 40,
+          left: 0,
+          height: 1,
+          width: "100%",
+          backgroundColor: "#e8e7e7",
         },
-        "& .rs-zoomslider-wrapper": {
-          margin: 0,
-        },
-        "& .ol-zoomslider": {
-          border: "1px solid #5a5a5a !important",
-          paddingBottom: "3px !important",
-        },
-        "& .ol-zoomslider-thumb": {
-          background: `${theme.colors.darkGray} !important`,
-        },
+        "&:hover": { boxShadow: "0 0 12px 2px rgb(0 0 0 / 90%)" },
         "& .rs-zoom-in, .rs-zoom-out": {
           background: "white",
-          boxShadow: "0 0 7px rgb(0 0 0 / 90%)",
-          borderRadius: "50%",
-          height: "40px",
-          width: "40px",
+          height: 40,
+          width: 40,
           display: "flex",
           justifyContent: "center",
           alignItems: "center",
-          transition: "box-shadow 0.5s ease",
           border: "none",
+          color: theme.colors.darkGray,
           "& svg": {
             "& path": {
               transition: "stroke 0.5s ease",
-              fill: theme.colors.darkGray,
+              fill: "currentColor",
+              stroke: "currentColor",
             },
           },
-          "&:hover": { boxShadow: "0 0 12px 2px rgb(0 0 0 / 90%)" },
-          "&:disabled": {
-            boxShadow: "0 0 7px 2px rgb(0 0 0 / 40%)",
-            cursor: "default",
+          "&:hover": {
+            color: theme.palette.secondary.dark,
           },
+        },
+        "& .rs-zoom-in": {
+          borderRadius: "20px 20px 0 0",
+        },
+        "& .rs-zoom-out": {
+          borderRadius: "0 0 20px 20px",
         },
       },
     },
@@ -83,13 +86,14 @@ const useStyles = makeStyles((theme) => {
 function MapControls({
   menuToggler = false,
   geolocation = true,
-  zoomSlider = true,
   fitExtent = true,
+  floorSwitcher = false,
   children,
 }) {
   const dispatch = useDispatch();
   const overlayWidth = useOverlayWidth();
   const screenHeight = useSelector((state) => state.app.screenHeight);
+  const activeTopic = useSelector((state) => state.app.activeTopic);
   const isSmallHeight = useMemo(() => {
     return ["xs", "s"].includes(screenHeight);
   }, [screenHeight]);
@@ -101,64 +105,18 @@ function MapControls({
   });
   const { t } = useTranslation();
   const map = useSelector((state) => state.app.map);
-  const [zoomSliderRef, setZoomSliderRef] = useState(null);
-
-  useEffect(() => {
-    let key = null;
-    // on resize reload the zoomSlider control
-    if (zoomSlider) {
-      key = map.on("change:size", () => {
-        const control = map
-          .getControls()
-          .getArray()
-          .find((ctrl) => ctrl instanceof ZoomSlider);
-        if (control) {
-          // Force reinitialization of the control
-          control.sliderInitialized_ = false;
-          setZoomSliderRef(control.target_);
-        } else {
-          setZoomSliderRef(null);
-        }
-      });
-    }
-    return () => {
-      if (map && key) {
-        unByKey([key]);
-      }
-    };
-  }, [dispatch, map, zoomSlider]);
-
-  useEffect(() => {
-    const onZoomSliderRefUpdate = () => {
-      dispatch(setZoomType("slider"));
-    };
-    if (zoomSliderRef) {
-      zoomSliderRef?.firstChild?.addEventListener(
-        "mousedown",
-        onZoomSliderRefUpdate,
-      );
-    }
-    return () => {
-      if (zoomSliderRef) {
-        zoomSliderRef?.firstChild?.removeEventListener(
-          "mousedown",
-          onZoomSliderRefUpdate,
-        );
-      }
-    };
-  }, [dispatch, zoomSliderRef]);
 
   return (
     <div
       className={`wkp-map-controls ${classes.mapControls}`}
       data-testid="map-controls-wrapper"
     >
-      {menuToggler && <MenuToggler />}
+      {menuToggler && (activeTopic?.menuToggler ?? <MenuToggler />)}
       <Zoom
         map={map}
         zoomInChildren={<ZoomIn />}
         zoomOutChildren={<ZoomOut />}
-        zoomSlider={!isSmallHeight && !isMobile && zoomSlider}
+        zoomSlider={false}
         titles={{
           zoomIn: t("Hineinzoomen"),
           zoomOut: t("Rauszoomen"),
@@ -172,6 +130,7 @@ function MapControls({
       />
       {geolocation && <Geolocation />}
       {fitExtent && <FitExtent />}
+      {floorSwitcher && <FloorSwitcher />}
       {children}
     </div>
   );